IMPLEMENTATION / added the tasks - user relationships (#9)
* IMPLEMENTATION / added the tasks - user relationships

* added full relationship for tasks - users //  fixed the user data leakage

+++ src/tasks/tasks.service.ts
@@ -4,7 +4,7 @@
  * @Last Modified by: Ardrit Krasniqi ©
  * @Last Modified time: 2021-12-10 23:23:36
  */
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 
 import { CreateTaskDto } from './dto/create-task-dto';
 import { GetTasksFilterDto } from './dto/get-tasks-filter-dto';
@@ -12,6 +12,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/users/user.entity';
 
 
 @Injectable()
@@ -23,35 +24,45 @@ export class TasksService {
     //TODO get the tasks from the db 
 
    
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]>{
+    async getTasks(
+        filterDto: GetTasksFilterDto,
+        user: User
+        ): Promise<Task[]>{
         //get tasks either filtered or all
-        const tasks = await this.taskRepository.getTasks(filterDto);
-
-
+        const tasks = await this.taskRepository.getTasks(filterDto, user);
         return tasks;
     }
 
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOneOrFail(id);
-        return found;
+    async getTaskById(
+        id: number, 
+        user: User
+        ): Promise<Task> {
+        const task = await this.taskRepository.findOne({where: {id, userId: user.id}});
+        if(!task){
+            throw new NotFoundException(`Task with ID "${id}" not found`)
+        }
+        return task;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return await this.taskRepository.createTask(createTaskDto);
+    async createTask(
+        createTaskDto: CreateTaskDto,
+        user: User
+        ): Promise<Task> {
+        return await this.taskRepository.createTask(createTaskDto, user);
     }
 
 
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id});
         if (!result.affected){
-            throw new NotFoundException(`Task with ID ${id} does not exist!`);
+            throw new NotFoundException(`Task with ID "${id}" does not exist!`);
         } 
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, user: User,  status: TaskStatus): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
